Problem: testing if anbud got jest going in devops
Solution: push a jest test see if it fails.
