+++ imports/api/auctions/auctions.js
@@ -1,3 +1,30 @@
-import { Mongo } from 'meteor/mongo'
+import { Mongo } from 'meteor/mongo';
+import { developmentValidationEnabledFalse } from '../indexDB'
+import SimpleSchema from 'simpl-schema';
 
 export const Auctions = new Mongo.Collection('auctions')
+
+var { Integer, RegEx } = SimpleSchema
+var { Id, Domain } = RegEx
+
+Auctions.schema = new SimpleSchema({
+  _id: { type: Id },
+  name: { type: String },
+  description: { type: String },
+  options: { type: Object },
+  "options.amount": { type: Integer },
+  "options.baseCurrency": { type: String },
+  "options.acceptedCurrency": { type: String },
+  "options.timeout": { type: Integer },
+  "options.reserve": { type: Number },
+  "options.reserveMet": {type: Boolean},
+  "options.highest": {type: Number},
+  createdBy: {type: String},
+  createdAt: {type: String}
+}, { requiredByDefault: developmentValidationEnabledFalse });
+
+Auctions.deny({
+  insert() { return true; },
+  update() { return true; },
+  remove() { return true; },
+});
