+++ imports/api/ratings/methods.js
@@ -1,7 +1,11 @@
 import { Meteor } from 'meteor/meteor'
 import { Ratings, RatingsTemplates, UserData, WalletImages } from '/imports/api/indexDB.js'
-import { log } from '/server/main.js'
 import { creditUserWith, removeUserCredit } from '/imports/api/utilities.js'
+if (Meteor.isServer){
+	import { log } from '/server/main.js'
+} else {
+	const log = {error(){}}
+}
 
 Meteor.methods({
 	answerRating: function(ratingId, winner) {
