+++ imports/api/wallet/methods.js
@@ -1,8 +1,12 @@
 import { Meteor } from 'meteor/meteor'
 import { Wallet, WalletImages, Currencies, Ratings, Bounties, REWARDCOEFFICIENT } from '/imports/api/indexDB.js'
-import { log } from '/server/main.js'
 import { creditUserWith, removeUserCredit } from '/imports/api/utilities.js'
-import { sendMessage } from '/imports/api/activityLog/server/methods'
+import { sendMessage } from '/imports/api/activityLog/methods'
+if (Meteor.isServer){
+	import { log } from '/server/main.js'
+} else {
+	const log = {error(){}}
+}
 
 Meteor.methods({
 	initializeWallet: function() {
