+++ imports/ui/components/typeahead.html
@@ -3,3 +3,7 @@
     <input class="typeahead form-control" id={{id}} type="text" placeholder={{placeholder}} style="width:15em">
   </div>
 </template>
+
+<template name="typeaheadSnapshotTest">
+  {{> typeahead params}}
+</template>
