+++ imports/ui/pages/bounties/bounties.js
@@ -2,14 +2,16 @@ import { Template } from 'meteor/templating';
 import { Bounties, REWARDCOEFFICIENT, Problems, Currencies } from '/imports/api/indexDB.js';
 import Cookies from 'js-cookie';
 
-export function calculateReward (now) { //used in bountyRender too
+function calculateReward (now) { //used in bountyRender too
     let add = 0
     if (this._id === 'new-currency') {
         add = 1
     }
-
-    return (this.credit ? this.credit.reduce((i1, i2) => i1 + i2.bounty, 0) : (((now - this.creationTime) / REWARDCOEFFICIENT) * (this.multiplier || 1) + add)).toFixed(6)
+    var res = (this.credit ? this.credit.reduce((i1, i2) => i1 + i2.bounty, 0) : (((now - this.creationTime) / REWARDCOEFFICIENT) * (this.multiplier || 1) + add))
+    return res
 }
+
+export const LocalBounties = new Mongo.Collection(null)
 
 import './bounties.html'
 import './rainbow.js';
@@ -34,7 +36,7 @@ Template.bounties.onCreated(function(){
   this.times = new ReactiveVar({})
   this.bounties = new ReactiveVar(null)
 
-  this.LocalBounties = new Mongo.Collection(null);
+  this.LocalBounties = LocalBounties
   this.currentIds = new ReactiveVar(null)
 
   this.filter = new ReactiveVar({})
@@ -70,6 +72,7 @@ Template.bounties.onCreated(function(){
     this.times.set(times)
   })
 
+   //  references to LocalBounty have to be nonreactive since rewards are appended to documents separately causing a loop with reactivity on
   this.autorun((comp)=>{
     let problems = Problems.find({ // dont show questions in here
       $or: [{
@@ -135,22 +138,14 @@ Template.bounties.onCreated(function(){
       pendingApproval: false
     }).fetch().map(i => {
       i.creationTime = i.creationTime || Template.instance().times.get()[i._id]
-
       return i
-    }), problems, currencies).sort((i1, i2) => {
-      return calculateReward.call(i2, Session.get('now'), Template.instance()) - calculateReward.call(i1, Session.get('now'))
-    })
+    }), problems, currencies)
 
     //this map copies over values to local collection so that updates can be diffed in order to notify user that bounty reward has been reset
     //  and adds last reward from when bounty was completed previously
+    //  has to be nonreactive since rewards are appended to documents separately causing a loop with reactivity on
+    Tracker.nonreactive(()=>{
     union.map((x, i)=>{
-      // return this.LocalBounties.insert(x)
-      //note the field projection
-      // var lastCompletedBounty = Bounties.findOne({type: x._id, completed: true, currentReward: {$exists: true}}, {sort: {completedAt: -1, expiresAt: -1}, fields: {currentReward: 1, completedAt: 1, currentUsername: 1}})
-      // last bounty get deleted when the user is rewarded, so it's not a working solution
-
-      //preserves previous sort operation
-      x.sort = i
       var copy = this.LocalBounties.findOne(x._id)
       if (copy) {
         this.LocalBounties.update(x._id, x)
@@ -158,12 +153,21 @@ Template.bounties.onCreated(function(){
         this.LocalBounties.insert(x)
       }
     })
+  })
 
       //lets filter out removed/replaced items from union since _id field is transcient for Bounties
       var ids = union.map(x=>x._id)
       this.currentIds.set(ids)
-
   })
+
+  this.autorun(()=>{
+    this.LocalBounties.find({_id: {$in: this.currentIds.get()}}).forEach((x, i)=>{
+      this.LocalBounties.update(x._id, {$set: { reward: calculateReward.call(x, Session.get('now'))}})
+    })
+  })
+
+  window.calculateReward = calculateReward
+  window.LocalBounties = this.LocalBounties
 })
 
 Template.bounties.onRendered(function(){
@@ -236,8 +240,11 @@ Template.bounties.helpers({
 		//returns transformed group of collections
 		filter = _.extend(templ.filter.get(), {
           _id: { $in: templ.currentIds.get() }, currentUserId: { $ne: Meteor.userId() ? Meteor.userId() : "" }
-		})
-		return templ.LocalBounties.find(filter, {sort: {sort: 1}})
+    })
+    //avoids rerendering whole list if reward is changed which it does
+    return templ.LocalBounties.find(filter, 
+      {sort: {reward: -1}, fields: {reward: 0}}
+    )
 	},
 	currentActiveBounties: function () {
 		var templ = Template.instance();
