+++ imports/ui/pages/compareCurrencies/compareCurrencies.html
@@ -1,16 +1,21 @@
 <template name="compareCurrencies">
+	<h3 class="mt-3">Compare currencies</h3>
     <div class="card border-secondary mb-3 mb-3" style="min-width:300px;overflow:hidden;">
         <div class="card-body">
             <div class="currencyDetailBox">
-                <div style="width:100%">
-                    <canvas class="mx-auto" id="radar"></canvas>
-                </div>
+				<div class="row">
+					<div class="col-sm-8 offset-sm-2">
+						<div style="width:100%">
+		                    <canvas id="radar" width="600" height="250"></canvas>
+		                </div>
+					</div>
+				</div>
                 <br /><br />
                 <div style="width:100%">
                     <canvas class="mx-auto" id="bar"></canvas>
                 </div>
                 <br />
-                <div id="js-comparing">
+                <div id="js-comparing" class="table-responsive">
                     <table class="table">
                         <thead>
                             <tr>
@@ -18,7 +23,14 @@
                                     {{>typeahead typeAheadProps}}
                                 </th>
                                 {{#each comparedCurrencies}}
-                                <th scope="col"><h3 style="color: {{color}};display: inline;"><a href="/currency/{{slug}}"><i class="fa fa-external-link"></i></a>{{currencyName}}<sup><a href="#" class="js-delete">X</a></sup></h3></th>
+                                <th scope="col">
+									<h5 style="color: {{color}}; ">
+										<a href="/currency/{{slug}}">
+											<i class="fa fa-external-link"></i>
+										</a>
+										{{currencyName}}<sup><a href="#" class="js-delete">X</a></sup>
+									</h5>
+								</th>
                                 {{/each}}
 
                             </tr>
