+++ imports/ui/pages/currencyDetail/feature.html
@@ -56,7 +56,10 @@
 		{{#each comments}}
 			{{> commentRender}}
 		{{else}}
-			{{> empty}}
+<div id="replyBox-{{_id}}">
+    <small for="replyButton" class="mt-3 ml-1">No messages yet, be the first</small>
+    <button class="btn btn-sm mb-2 ml-2 btn-primary pull-right comments" id="replyButton">Reply</button>
+</div>
 		{{/each}}
     {{else}}
 		{{> loading}}
