+++ imports/ui/pages/currencyDetail/feature.js
@@ -120,10 +120,13 @@ Template.feature.events({
   'focus .replyText': function() {
     $(".replyFooter-" + this._id).show();
   },
-  'click .comments': function() {
-    if(Cookies.get("submitted" + this._id) != "true") {
-    $(".newcomment-" + this._id).show();
-  };
+    'click .comments': function() {
+      if(Cookies.get("submitted" + this._id) != "true") {
+      $(".newcomment-" + this._id).show();
+      $("#replyBox-" + this._id).hide();
+      $("#replyText-" + this._id).focus();
+
+    };
   if(Template.instance().showingComments.get(this._id) != "true") {
     $(".commentParent-" + this._id).show();
     Template.instance().showingComments.set(this._id, "true")
