+++ imports/ui/pages/currencyDetail/redflag.html
@@ -48,7 +48,10 @@
 {{#each comments}}
   {{> redflagComment}}
 {{else}}
-  {{> empty}}
+  <div id="replyBox-{{_id}}">
+    <small for="replyButton" class="mt-3 ml-1">No messages yet, be the first</small>
+    <button class="btn btn-sm mb-2 ml-2 btn-primary pull-right comments" id="replyButton">Reply</button>
+</div>
 {{/each}}
 {{else}}
 {{> loading}}
