+++ jest.config.js
@@ -1,0 +1,18 @@
+
+// module.exports = {
+//   moduleFileExtensions: [
+//     'js',
+//   ],
+//   modulePaths: [
+//     '<rootDir>/node_modules/',
+//     '<rootDir>/node_modules/jest-meteor-stubs/lib/',
+//   ],
+//   moduleNameMapper: {
+//     "meteor/(.*)": "<rootDir>/tests/mocks/$1.js",
+//     '^(.*):(.*)$': '$1_$2',
+//   },
+//   unmockedModulePathPatterns: [
+//     '/^imports\\/.*\\.jsx?$/',
+//     '/^node_modules/',
+//   ],
+// };
